refactor(*): Reviewed configuration & misc cleanup

+++ swabbie-core/src/main/kotlin/com/netflix/spinnaker/swabbie/AbstractResourceTypeHandler.kt
@@ -16,15 +16,17 @@
 
 package com.netflix.spinnaker.swabbie
 
+import com.google.common.collect.Lists
 import com.netflix.spectator.api.Registry
 import com.netflix.spinnaker.kork.core.RetrySupport
 import com.netflix.spinnaker.kork.lock.LockManager.LockOptions
-import com.netflix.spinnaker.swabbie.echo.EchoService
-import com.netflix.spinnaker.swabbie.echo.Notifier
 import com.netflix.spinnaker.swabbie.events.*
 import com.netflix.spinnaker.swabbie.exclusions.ResourceExclusionPolicy
 import com.netflix.spinnaker.swabbie.exclusions.shouldExclude
 import com.netflix.spinnaker.swabbie.model.*
+import com.netflix.spinnaker.swabbie.notifications.Notifier
+import kotlinx.coroutines.experimental.launch
+import kotlinx.coroutines.experimental.runBlocking
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.context.ApplicationEventPublisher
@@ -41,7 +43,7 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
   private val resourceRepository: ResourceTrackingRepository,
   private val exclusionPolicies: List<ResourceExclusionPolicy>,
   private val ownerResolver: OwnerResolver<T>,
-  private val notifier: Notifier,
+  private val notifiers: List<Notifier>,
   private val applicationEventPublisher: ApplicationEventPublisher,
   private val lockingService: Optional<LockingService>,
   private val retrySupport: RetrySupport
@@ -80,15 +82,15 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
    */
   private fun execute(workConfiguration: WorkConfiguration, callback: () -> Unit, action: Action) {
     when (action) {
-      Action.MARK -> withLocking(workConfiguration) {
+      Action.MARK -> withLocking(action, workConfiguration) {
         doMark(workConfiguration, callback)
       }
 
-      Action.DELETE -> withLocking(workConfiguration) {
+      Action.DELETE -> withLocking(action, workConfiguration) {
         doDelete(workConfiguration, callback)
       }
 
-      Action.NOTIFY -> withLocking(workConfiguration) {
+      Action.NOTIFY -> withLocking(action, workConfiguration) {
         doNotify(workConfiguration, callback)
       }
 
@@ -96,15 +98,17 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
     }
   }
 
-  private fun withLocking(workConfiguration: WorkConfiguration,
-                          callback: () -> Unit
+  private fun withLocking(
+    action: Action,
+    workConfiguration: WorkConfiguration,
+    callback: () -> Unit
   ) {
     if (lockingService.isPresent) {
-      val normalizedName = workConfiguration.namespace
+      val normalizedLockName = ("${action.name}." + workConfiguration.namespace)
         .replace(":", ".")
-
+        .toLowerCase()
       val lockOptions = LockOptions()
-        .withLockName(normalizedName.toLowerCase())
+        .withLockName(normalizedLockName)
         .withMaximumLockDuration(lockingService.get().swabbieMaxLockDuration)
       lockingService.get().acquireLock(lockOptions, {
         callback.invoke()
@@ -117,7 +121,7 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
 
   private fun doMark(workConfiguration: WorkConfiguration, postMark: () -> Unit) {
     // initialize counters
-    excludedResourcesDuringMarkCounter.set(0)
+    exclusionCounters[Action.MARK] = AtomicInteger(0)
     val timerId = markDurationTimer.start()
     val candidateCounter = AtomicInteger(0)
     val violationCounter = AtomicInteger(0)
@@ -126,42 +130,45 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
     try {
       log.info("${javaClass.simpleName} running. Configuration: ", workConfiguration)
       val candidates: List<T>? = getCandidates(workConfiguration)
-      log.info("fetched {} resources. Configuration: {}", candidates?.size, workConfiguration)
       if (candidates == null || candidates.isEmpty()) {
         return
       }
 
+      log.info("fetched {} resources. Configuration: {}", candidates.size, workConfiguration)
       totalResourcesVisitedCounter.set(candidates.size)
-      candidates.forEach { candidate ->
-        candidate.set(resourceOwnerField, ownerResolver.resolve(candidate))
+      candidates.let {
+        resolveOwners(workConfiguration, it)
       }
 
-      val markedCandidates: List<MarkedResource> = Optional.ofNullable(resourceRepository.getMarkedResources())
-        .orElse(emptyList())
-        .filter {
-          it.namespace == workConfiguration.namespace
-        }
+      candidates.filterOutExcludedCandidates(workConfiguration).let { filteredCandidates ->
+        val maxItemsToProcess = Math.min(filteredCandidates.size, workConfiguration.maxItemsProcessedPerCycle)
 
-      // filter out resources that should be excluded
-      candidates.filter {
-        !shouldExcludeResource(it, workConfiguration, Action.MARK)
-      }.forEach { candidate ->
-        try {
-          getViolations(candidate).let { violations ->
+        // list of currently marked & stored resources
+        val markedCandidates: List<MarkedResource> = Optional.ofNullable(resourceRepository.getMarkedResources())
+          .orElse(emptyList())
+          .filter {
+            it.namespace == workConfiguration.namespace
+          }
+
+        for (candidate in filteredCandidates) {
+          if (candidateCounter.get() > maxItemsToProcess) {
+            log.info("Max items to process reached {}...", maxItemsToProcess)
+            break
+          }
+
+          try {
+            val violations: List<Summary> = candidate.getViolations()
             markedCandidates.find {
               it.resourceId == candidate.resourceId
             }.let { alreadyMarkedCandidate ->
               when {
-                alreadyMarkedCandidate != null && violations.isEmpty() -> ensureResourceUnmarked(
-                  alreadyMarkedCandidate,
-                  workConfiguration
-                )
-                alreadyMarkedCandidate == null && !violations.isEmpty() -> {
+                violations.isEmpty() -> ensureResourceUnmarked(alreadyMarkedCandidate, workConfiguration)
+                !violations.isEmpty() && alreadyMarkedCandidate == null -> {
                   val newMarkedResource = MarkedResource(
                     resource = candidate,
                     summaries = violations,
                     namespace = workConfiguration.namespace,
-                    resourceOwner = candidate.details[resourceOwnerField] as? String,
+                    resourceOwner = candidate.details[resourceOwnerField] as String,
                     projectedDeletionStamp = deletionTimestamp(workConfiguration)
                   )
 
@@ -177,13 +184,12 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
                 }
               }
             }
+          } catch (e: Exception) {
+            log.error("Failed while invoking ${javaClass.simpleName}", e)
+            recordFailureForAction(Action.MARK, workConfiguration, e)
           }
-        } catch (e: Exception) {
-          log.error("Failed while invoking ${javaClass.simpleName}", e)
-          recordFailureForAction(Action.MARK, workConfiguration, e)
         }
       }
-
       printResult(candidateCounter, totalResourcesVisitedCounter, workConfiguration, markedResourceIds, Action.MARK)
     } finally {
       recordMarkMetrics(timerId, workConfiguration, violationCounter, candidateCounter)
@@ -191,11 +197,31 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
     }
   }
 
+  private fun List<T>.filterOutExcludedCandidates(workConfiguration: WorkConfiguration): List<T> {
+    return this.filter {
+      !shouldExcludeResource(it, workConfiguration, Action.MARK)
+    }
+  }
+
+  private fun resolveOwners(
+    workConfiguration: WorkConfiguration,
+    candidates: List<T>
+  ) = runBlocking<Unit> {
+    candidates.forEach { candidate ->
+      launch {
+        candidate.set(
+          resourceOwnerField,
+          ownerResolver.resolve(candidate) ?: workConfiguration.notificationConfiguration.defaultDestination
+        )
+      }
+    }
+  }
+
   private fun ensureResourceUnmarked(
-    markedResource: MarkedResource,
+    markedResource: MarkedResource?,
     workConfiguration: WorkConfiguration
   ) {
-    if (!workConfiguration.dryRun) {
+    if (markedResource != null && !workConfiguration.dryRun) {
       log.info("{} is no longer a candidate.", markedResource)
       resourceRepository.remove(markedResource)
       applicationEventPublisher.publishEvent(UnMarkResourceEvent(markedResource, workConfiguration))
@@ -212,15 +238,15 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
     markedResourceIds: MutableMap<String, List<Summary>>,
     action: Action
   ) {
-    log.info("** ${action.name} Summary: {} CLEANUP CANDIDATES OUT OF {} SCANNED. EXCLUDED {}. CONFIGURATION: {}**",
+    log.info("** ${action.name} Summary: {} CANDIDATES OUT OF {} SCANNED. EXCLUDED {}. CONFIGURATION: {}**",
       candidateCounter.get(),
       totalResourcesVisitedCounter.get(),
-      excludedResourcesDuringMarkCounter.get(),
+      exclusionCounters[action],
       workConfiguration
     )
 
     if (candidateCounter.get() > 0) {
-      log.debug("** ${action.name} list Configuration: {} **", workConfiguration)
+      log.debug("** ${action.name} Result. Configuration: {} **", workConfiguration)
       markedResourceIds.forEach {
         log.debug("$it")
       }
@@ -233,23 +259,21 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
     action: Action? = null
   ): Boolean {
     val creationDate = Instant.ofEpochMilli(resource.createTs).atZone(ZoneId.systemDefault()).toLocalDate()
-    if (DAYS.between(creationDate, LocalDate.now()) < workConfiguration.maxAge) {
+    if ((action != null && action != Action.NOTIFY) &&
+      DAYS.between(creationDate, LocalDate.now()) < workConfiguration.maxAge) {
       log.debug("Excluding {} newer than {} days", resource, workConfiguration.maxAge)
       return true
     }
 
     return shouldExclude(resource, workConfiguration, exclusionPolicies, log).also { excluded ->
       if (excluded) {
-        if (action == Action.MARK) {
-          excludedResourcesDuringMarkCounter.incrementAndGet()
-        } else if (action == Action.DELETE) {
-          excludedResourcesDuringDeleteCounter.incrementAndGet()
-        }
+        exclusionCounters[action]?.incrementAndGet()
       }
     }
   }
 
   private fun doDelete(workConfiguration: WorkConfiguration, postClean: () -> Unit) {
+    exclusionCounters[Action.DELETE] = AtomicInteger(0)
     val candidateCounter = AtomicInteger(0)
     val markedResourceIds = mutableMapOf<String, List<Summary>>()
     try {
@@ -259,7 +283,13 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
         return
       }
 
+      val maxItemsToProcess = Math.min(toDelete.size, workConfiguration.maxItemsProcessedPerCycle)
       for (r in toDelete) {
+        if (candidateCounter.get() > maxItemsToProcess) {
+          log.info("Max items to process reached {}...", maxItemsToProcess)
+          break
+        }
+
         try {
           val candidate: T? = getCandidate(r, workConfiguration)
           if (candidate == null) {
@@ -273,14 +303,14 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
             continue
           }
 
-          getViolations(candidate).let { violations ->
+          candidate.getViolations().let { violations ->
             if (violations.isEmpty()) {
               ensureResourceUnmarked(r, workConfiguration)
             } else {
               if (r.notificationInfo != null && !workConfiguration.dryRun) {
                 retrySupport.retry({
                   deleteMarkedResource(r, workConfiguration)
-                }, 3, 5000, true)
+                }, maxAttempts, timeoutMillis, true)
 
                 applicationEventPublisher.publishEvent(DeleteResourceEvent(r, workConfiguration))
                 resourceRepository.remove(r)
@@ -303,82 +333,108 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
     }
   }
 
-  private fun getViolations(candidate: T): List<Summary> {
+  private fun T.getViolations(): List<Summary> {
     return rules.mapNotNull {
-      it.apply(candidate).summary
+      it.apply(this).summary
     }
   }
 
   private fun doNotify(workConfiguration: WorkConfiguration, postNotify: () -> Unit) {
+    exclusionCounters[Action.NOTIFY] = AtomicInteger(0)
+    val candidateCounter = AtomicInteger(0)
     try {
-      log.info("Notification Agent Started with configuration {}", workConfiguration)
-      val owners = mutableMapOf<String, MutableList<MarkedResource>>()
-      resourceRepository.getMarkedResources()
-        ?.filter {
-          workConfiguration.namespace.equals(it.namespace, ignoreCase = true) && it.notificationInfo == null
-        }?.forEach { markedResource ->
-          markedResource.resourceOwner?.let { owner ->
-            if (owners[owner] == null) {
-              owners[owner] = mutableListOf(markedResource)
-            } else {
-              owners[owner]!!.add(markedResource)
-            }
-          }
-        }
-
-      owners.let { ownersToResources ->
-        ownersToResources.forEach { ownerToResources ->
-          log.info("DryRun={}, shouldNotify={}, user={}, {} cleanup candidate(s)",
-            workConfiguration.dryRun,
-            workConfiguration.notificationConfiguration!!.notifyOwner,
-            ownerToResources.key,
-            ownerToResources.value.size
-          )
-
-          if (workConfiguration.dryRun || !workConfiguration.notificationConfiguration.notifyOwner) {
-            log.info("Skipping notifications. Configuration: {}", workConfiguration)
-          } else {
-            ownerToResources.value.let { resources ->
-              notifier.notify(
-                recipient = ownerToResources.key,
-                messageType = EchoService.Notification.Type.EMAIL.name,
-                additionalContext = mapOf(
-                  "resourceOwner" to ownerToResources.key,
-                  "resources" to resources,
-                  "configuration" to workConfiguration,
-                  "resourceType" to workConfiguration.resourceType.formatted(),
-                  "spinnakerLink" to workConfiguration.notificationConfiguration.spinnakerResourceUrl,
-                  "optOutLink" to workConfiguration.notificationConfiguration.optOutUrl
-                )
-              )
-
-              log.debug("Notification sent to {} for {}", ownerToResources.key, resources)
-              resources.forEach { resource ->
-                val offsetStampSinceMarked: Long = ChronoUnit.MILLIS.between(
-                  Instant.ofEpochMilli(resource.createdTs!!),
-                  Instant.now(clock)
-                )
-
-                resource.apply {
-                  projectedDeletionStamp += offsetStampSinceMarked
-                  notificationInfo = NotificationInfo(
-                    recipient = ownerToResources.key,
-                    notificationStamp = Instant.now(clock).toEpochMilli(),
-                    notificationType = EchoService.Notification.Type.EMAIL.name
-                  )
-                }.also {
-                  resourceRepository.upsert(it)
-                  applicationEventPublisher.publishEvent(OwnerNotifiedEvent(resource, workConfiguration))
-                }
-              }
-            }
-          }
-        }
+      if (workConfiguration.dryRun || !workConfiguration.notificationConfiguration.enabled) {
+        log.info("Notification not enabled for {}. Skipping...", workConfiguration)
+        return
       }
-    } catch (e: Exception) {
-      recordFailureForAction(Action.NOTIFY, workConfiguration, e)
+
+      val markedResources = resourceRepository.getMarkedResources()
+      if (markedResources == null || markedResources.isEmpty()) {
+        log.info("Nothing to notify. Skipping...", workConfiguration)
+        return
+      }
+
+      val markedResourceIds = mutableMapOf<String, List<Summary>>()
+      val maxItemsToProcess = Math.min(markedResources.size, workConfiguration.maxItemsProcessedPerCycle)
+      markedResources.subList(0, maxItemsToProcess)
+        .filter {
+          workConfiguration.namespace == it.namespace &&
+            !shouldExcludeResource(it.resource as T, workConfiguration, Action.NOTIFY) && it.notificationInfo == null
+        }.groupBy {
+          it.resourceOwner
+        }.forEach { ownersAndResources ->
+          Lists.partition(
+            ownersAndResources.value,
+            workConfiguration.notificationConfiguration.itemsPerMessage
+          ).forEach { partition ->
+            try {
+              sendNotification(ownersAndResources.key, workConfiguration, partition)
+              partition.forEach { resource ->
+                val offsetStampSinceMarked = ChronoUnit.MILLIS
+                  .between(Instant.ofEpochMilli(resource.markTs!!), Instant.now(clock))
+
+                resource
+                  .apply {
+                    projectedDeletionStamp += offsetStampSinceMarked
+                    notificationInfo = NotificationInfo(
+                      recipient = ownersAndResources.key,
+                      notificationStamp = Instant.now(clock).toEpochMilli(),
+                      notificationType = Notifier.NotificationType.EMAIL.name
+                    )
+                  }.also {
+                    resourceRepository.upsert(it)
+                    applicationEventPublisher.publishEvent(OwnerNotifiedEvent(resource, workConfiguration))
+                    candidateCounter.set(partition.size)
+                  }
+
+                markedResourceIds[resource.resourceId] = resource.summaries
+              }
+              log.debug("Notification sent to {} for {} resources", ownersAndResources.key, partition.size)
+            } catch (e: Exception) {
+              recordFailureForAction(Action.NOTIFY, workConfiguration, e)
+            }
+          }
+        }
+
+      printResult(
+        candidateCounter,
+        AtomicInteger(markedResources.size),
+        workConfiguration,
+        markedResourceIds,
+        Action.NOTIFY
+      )
     } finally {
       postNotify.invoke()
+    }
+  }
+
+  private fun sendNotification(
+    owner: String,
+    workConfiguration:
+    WorkConfiguration,
+    resources: List<MarkedResource>
+  ) {
+    notifiers.forEach { notifier ->
+      workConfiguration.notificationConfiguration.types.forEach { notificationType ->
+        if (notificationType.equals(Notifier.NotificationType.EMAIL.name, true)) {
+          val notificationContext = mapOf(
+            "resourceOwner" to owner,
+            "resources" to resources,
+            "configuration" to workConfiguration,
+            "resourceType" to workConfiguration.resourceType.formatted(),
+            "spinnakerLink" to workConfiguration.notificationConfiguration.resourceUrl,
+            "optOutLink" to workConfiguration.notificationConfiguration.optOutBaseUrl
+          )
+
+          retrySupport.retry({
+            notifier.notify(
+              recipient = owner,
+              messageType = notificationType,
+              additionalContext = notificationContext
+            )
+          }, maxAttempts, timeoutMillis, true)
+        }
+      }
     }
   }
 
@@ -388,3 +444,6 @@ abstract class AbstractResourceTypeHandler<out T : Resource>(
   private val Period.fromNow: LocalDate
     get() = LocalDate.now(clock) + this
 }
+
+const val timeoutMillis: Long = 5000
+const val maxAttempts: Int = 3
